+++ src/csv_reader.py
@@ -1,3 +1,7 @@
+#
+# 2025-01-19
+#
+
 import csv
 import logger
 
@@ -5,42 +9,49 @@ from text_grid import TextGrid
 
 # -----------------------------------------------------------------------------
 
+def __check_row_valid(row_cells: list[str]) -> bool:
+    # ignore rows with empty cells 'A,B,C' or cell 'A' with a long horizontal line
+    row_valid = (len(row_cells) > 3) and (row_cells[0] or row_cells[1] or row_cells[2])
+    row_valid = row_valid and not row_cells[0].startswith("___")
+    return row_valid
+
 def __read_sp(rows: list[str], tg: TextGrid):
     max_cols = 0
     for row in rows:
         # split row by any number of following whitespaces
         row_cells = row.split()
-        max_cols = max(max_cols, len(row_cells))
-        # ignore rows with empty cell 'A' or cell with a long horizontal line
-        if len(row_cells) > 0 and row_cells[0] != "" and not row_cells[0].startswith("___"):
-            row_cells_processed = []
+        if not __check_row_valid(row_cells):
+            break
 
-            # merge quoted cells into single one,
-            # like this: "5k1 5% 0603"
-            quoted_cell = ""
-            for cell in row_cells:
-                if cell.startswith('"'):
-                    quoted_cell = cell
-                elif len(quoted_cell) > 0:
-                    quoted_cell += ' '
-                    quoted_cell += cell
-                    if cell.endswith('"'):
-                        # drop the quotes
-                        quoted_cell = quoted_cell[1:-1]
-                        row_cells_processed.append(quoted_cell)
-                        quoted_cell = ""
-                else:
-                    row_cells_processed.append(cell.strip())
-            tg.rows_raw().append(row_cells_processed)
+        max_cols = max(max_cols, len(row_cells))
+        row_cells_processed = []
+        # merge quoted cells into single one,
+        # like this: "5k1 5% 0603"
+        quoted_cell = ""
+
+        for cell in row_cells:
+            if cell.startswith('"'):
+                quoted_cell = cell
+            elif len(quoted_cell) > 0:
+                quoted_cell += ' '
+                quoted_cell += cell
+                if cell.endswith('"'):
+                    # drop the quotes
+                    quoted_cell = quoted_cell[1:-1]
+                    row_cells_processed.append(quoted_cell)
+                    quoted_cell = ""
+            else:
+                row_cells_processed.append(cell.strip())
+
+        tg.rows_raw().append(row_cells_processed)
     return max_cols
 
 def __read_csv(file, tg: TextGrid, delim: str, quote_char: str = '"'):
     max_cols = 0
     reader = csv.reader(file, delimiter=delim, quotechar=quote_char)
-    for row in reader:
-        # ignore rows with empty cell 'A' or cell with a long horizontal line
-        if len(row) > 0 and row[0] != "" and not row[0].startswith("___"):
-            row_cells = [cell.strip() for cell in row]
+    for row_cells in reader:
+        if __check_row_valid(row_cells):
+            row_cells = [cell.strip() for cell in row_cells]
             max_cols = max(max_cols, len(row_cells))
             tg.rows_raw().append(row_cells)
 
@@ -75,7 +86,7 @@ def read_csv(path: str, delim: str) -> TextGrid:
     tg = TextGrid()
     max_cols = 0
 
-    with open(path, 'r', encoding="utf-8") as f:
+    with open(path, "r", encoding="utf-8") as f:
         if delim == "*sp":
             rows = f.read().splitlines()
             max_cols = __read_sp(rows, tg)