+++ src/ods_reader.py
@@ -1,3 +1,7 @@
+#
+# 2025-01-19
+#
+
 import logger
 
 from odf import opendocument, table
@@ -11,6 +15,12 @@ from odf import opendocument, table
 from text_grid import TextGrid
 
 # -----------------------------------------------------------------------------
+
+def __check_row_valid(row_cells: list[str]) -> bool:
+    # ignore rows with empty cells 'A,B,C' or cell 'A' with a long horizontal line
+    row_valid = (len(row_cells) > 3) and (row_cells[0] or row_cells[1] or row_cells[2])
+    row_valid = row_valid and not row_cells[0].startswith("___")
+    return row_valid
 
 def read_ods_sheet(path: str) -> TextGrid:
     """
@@ -34,18 +44,19 @@ def read_ods_sheet(path: str) -> TextGrid:
             for tablerow in tab.getElementsByType(table.TableRow):
                 tablecells = tablerow.getElementsByType(table.TableCell)
                 row_cells = []
+
                 # when iterating through the row cells, take the "repeat" attribute into account
                 for cell in tablecells:
                     repeated = cell.getAttribute(REPEATS_ATTR) or 1
                     repeated = int(repeated)
                     for _ in range(repeated):
-                        row_cells.append(str(cell))
+                        row_cells.append(str(cell).strip())
+
+                if not __check_row_valid(row_cells):
+                    break
 
                 max_cols = max(max_cols, len(row_cells))
-
-                # ignore rows with empty cell 'A'
-                if row_cells and row_cells[0] != "":
-                    tg.rows_raw().append(row_cells)
+                tg.rows_raw().append(row_cells)
 
             tg.nrows = len(tg.rows_raw())
             tg.ncols = max_cols