+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
+## 1.3.1 - 2025-02-10
+
+* Added
+* Changed
+* Deprecated
+* Removed
+* Fixed
+  * corrected loading of the WiP files, after component idx was edded to the PnP editor
+
 ## 1.3.0 - 2025-01-18
 
 * Added